@@ -1,31 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-
+    <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
-
-    <Copyright>© 2019-2021 Doxa Labs Ltd. All Rights Reserved.</Copyright>
-
+    <Copyright>© 2019-2022 Doxa Labs Ltd. All Rights Reserved.</Copyright>
     <Authors>Doxa Labs Ltd</Authors>
-
     <Product>Doxa Lab's Excel Tools</Product>
-
     <Description>This tool provides a C# based solution to create Excel files Simple and Fast. This package supports Android, iOS, Linux, macOS and Windows.</Description>
-
-    <Version>2.1.0</Version>
-
+    <Version>2.2.0</Version>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-
     <PackageProjectUrl>https://doxalabs.co.uk</PackageProjectUrl>
-
     <RepositoryUrl>https://github.com/doxa-labs/ExcelLabs.git</RepositoryUrl>
-
     <RepositoryType>git</RepositoryType>
-
     <PackageTags>Excel Office xls xlsx IO C# OpenXml Dataset .NETStandard</PackageTags>
-
-    <PackageReleaseNotes>.NET 5 &amp; .NET Framework Demo</PackageReleaseNotes>
+    <PackageReleaseNotes>.NET 6 &amp; .NET Framework Demo</PackageReleaseNotes>
+    <AssemblyVersion>2.2.0</AssemblyVersion>
+    <FileVersion>2.2.0</FileVersion>
   </PropertyGroup>
 
   <ItemGroup>
